requirements.txt
celery==4.4.2
Django>=3.0,<4.0
django-dotenv==1.4.2
djangorestframework_simplejwt==5.2.0
django_rest_swagger==2.2.0
drf-yasg==1.20.0
psycopg2>=2.8
python-dateutil

